Move various build graph-related files to new pkg: build_graph.
pants/base was getting crowded. Moving this will make it somewhat
easier to later refactor/reimplement aspects of BUILD file parsing.

Note: doesn't move some core files like target and payload, as those
are heavily referenced in plugins.  A followup change *may* move
those, but with a proper deprecation cycle.

After this change there are no dependencies from base to build_graph.
Target does take a build_graph argument (but doesn't reference
the type explicitly).

Also moves cache_manager/build_invalidator to a new pants/invalidation
package, as this helped untangle weird deps.

Note to Foursquare people: This will break a couple of our internal
plugins, which will need to move their imports of VersionedTargetSet.

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/82646618.

Reviewed at https://rbcommons.com/s/twitter/r/2899/
