+++ src/python/pants/backend/codegen/tasks/simple_codegen_task.py
@@ -15,10 +15,10 @@ from pants.backend.core.tasks.task import Task
 from pants.base.address import Address
 from pants.base.address_lookup_error import AddressLookupError
 from pants.base.build_environment import get_buildroot
-from pants.base.build_graph import sort_targets
 from pants.base.dep_lookup_error import DepLookupError
 from pants.base.exceptions import TaskError
 from pants.base.workunit import WorkUnitLabel
+from pants.build_graph.build_graph import sort_targets
 from pants.util.dirutil import safe_rmtree, safe_walk
 from pants.util.memo import memoized_property
 from pants.util.meta import AbstractClass
