+++ src/python/pants/backend/core/tasks/group_task.py
@@ -11,8 +11,8 @@ from collections import defaultdict, deque
 from contextlib import contextmanager
 
 from pants.backend.core.tasks.task import Task, TaskBase
-from pants.base.build_graph import invert_dependencies
 from pants.base.workunit import WorkUnit, WorkUnitLabel
+from pants.build_graph.build_graph import invert_dependencies
 from pants.goal.goal import Goal
