+++ src/python/pants/backend/core/tasks/what_changed.py
@@ -10,7 +10,7 @@ import re
 from pants.backend.core.tasks.console_task import ConsoleTask
 from pants.base.build_environment import get_scm
 from pants.base.exceptions import TaskError
-from pants.base.source_mapper import SpecSourceMapper
+from pants.build_graph.source_mapper import SpecSourceMapper
 from pants.goal.workspace import ScmWorkspace
