+++ src/python/pants/backend/jvm/tasks/ivy_resolve.py
@@ -15,9 +15,8 @@ from pants.backend.jvm.targets.jar_dependency import JarDependency
 from pants.backend.jvm.tasks.classpath_products import ClasspathProducts
 from pants.backend.jvm.tasks.ivy_task_mixin import IvyTaskMixin
 from pants.backend.jvm.tasks.nailgun_task import NailgunTask
-from pants.base.cache_manager import VersionedTargetSet
-from pants.base.exceptions import TaskError
 from pants.binaries import binary_util
+from pants.invalidation.cache_manager import VersionedTargetSet
 from pants.util.dirutil import safe_mkdir
 from pants.util.memo import memoized_property
 from pants.util.strutil import safe_shlex_split
