+++ src/python/pants/backend/project_info/tasks/export.py
@@ -22,7 +22,6 @@ from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.backend.jvm.targets.jvm_app import JvmApp
 from pants.backend.jvm.targets.jvm_target import JvmTarget
 from pants.backend.jvm.targets.scala_library import ScalaLibrary
-from pants.backend.jvm.tasks.classpath_products import ArtifactClasspathEntry
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.backend.python.targets.python_target import PythonTarget
 from pants.backend.python.tasks.python_task import PythonTask
@@ -32,8 +31,8 @@ from pants.java.distribution.distribution import DistributionLocator
 from pants.util.memo import memoized_property
 
 
-# Changing the behavior of this task may affect the IntelliJ Pants plugin
-# Please add fkorotkov, tdesai to reviews for this file
+# Changing the behavior of this task may affect the IntelliJ Pants plugin.
+# Please add fkorotkov, tdesai to reviews for this file.
 class Export(PythonTask, ConsoleTask):
   """Generates a JSON description of the targets as configured in pants.
