+++ src/python/pants/backend/python/BUILD
@@ -104,7 +104,7 @@ python_library(
     'src/python/pants/backend/core/targets:common',
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/base:build_environment',
-    'src/python/pants/base:build_invalidator',
+    'src/python/pants/invalidation',
     'src/python/pants/util:dirutil'
   ],
 )
