+++ src/python/pants/backend/python/tasks/BUILD
@@ -30,7 +30,7 @@ python_library(
     'src/python/pants/backend/python:thrift_builder',
     'src/python/pants/base:address_lookup_error',
     'src/python/pants/base:build_environment',
-    'src/python/pants/base:build_graph',
+    'src/python/pants/build_graph',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:generator',
     'src/python/pants/base:hash_utils',
