+++ src/python/pants/goal/BUILD
@@ -33,7 +33,7 @@ python_library(
     ':products',
     ':workspace',
     'src/python/pants/base:build_environment',
-    'src/python/pants/base:build_graph',
+    'src/python/pants/build_graph',
     'src/python/pants/base:source_root',
     'src/python/pants/base:target', # XXX(fixme)
     'src/python/pants/base:worker_pool',
