+++ src/python/pants/invalidation/BUILD
@@ -1,0 +1,15 @@
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+
+python_library(
+  name = 'invalidation',
+  sources = globs('*.py'),
+  dependencies = [
+    'src/python/pants/base:hash_utils',
+    'src/python/pants/base:target',
+    'src/python/pants/build_graph',
+    'src/python/pants/fs',
+    'src/python/pants/util:dirutil',
+  ],
+)
