+++ tests/python/pants_test/backend/jvm/tasks/BUILD
@@ -163,8 +163,8 @@ python_tests(
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/tasks:ivy_resolve',
+    'src/python/pants/invalidation',
     'src/python/pants/backend/jvm:ivy_utils',
-    'src/python/pants/base:cache_manager',
     'src/python/pants/util:contextutil',
     'tests/python/pants_test/jvm:jvm_tool_task_test_base',
   ]
