+++ tests/python/pants_test/base_test.py
@@ -13,17 +13,17 @@ from tempfile import mkdtemp
 from textwrap import dedent
 
 from pants.base.address import Address
-from pants.base.build_configuration import BuildConfiguration
 from pants.base.build_file import FilesystemBuildFile
-from pants.base.build_file_address_mapper import BuildFileAddressMapper
 from pants.base.build_file_aliases import BuildFileAliases
-from pants.base.build_file_parser import BuildFileParser
-from pants.base.build_graph import BuildGraph
 from pants.base.build_root import BuildRoot
 from pants.base.cmd_line_spec_parser import CmdLineSpecParser
 from pants.base.exceptions import TaskError
 from pants.base.source_root import SourceRoot
 from pants.base.target import Target
+from pants.build_graph.build_configuration import BuildConfiguration
+from pants.build_graph.build_file_address_mapper import BuildFileAddressMapper
+from pants.build_graph.build_file_parser import BuildFileParser
+from pants.build_graph.build_graph import BuildGraph
 from pants.goal.goal import Goal
 from pants.goal.products import MultipleRootedProducts, UnionProducts
 from pants.subsystem.subsystem import Subsystem
