+++ tests/python/pants_test/build_graph/test_build_file_parser.py
@@ -12,8 +12,8 @@ from textwrap import dedent
 from pants.base.address import BuildFileAddress
 from pants.base.build_file import FilesystemBuildFile
 from pants.base.build_file_aliases import BuildFileAliases
-from pants.base.build_file_parser import BuildFileParser
 from pants.base.target import Target
+from pants.build_graph.build_file_parser import BuildFileParser
 from pants_test.base_test import BaseTest
