+++ tests/python/pants_test/build_graph/test_source_mapper.py
@@ -5,9 +5,10 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
+# TODO: Create a dummy target type in this test and remove this dep.
 from pants.backend.jvm.targets.java_library import JavaLibrary
 from pants.base.build_file_aliases import BuildFileAliases
-from pants.base.source_mapper import LazySourceMapper, SpecSourceMapper
+from pants.build_graph.source_mapper import LazySourceMapper, SpecSourceMapper
 from pants_test.base_test import BaseTest
