+++ tests/python/pants_test/cache/BUILD
@@ -4,18 +4,29 @@
 target(
   name = 'cache',
   dependencies = [
+    ':artifact',
     ':artifact_cache',
     ':cache_setup',
-    ':resolver',
+    ':resolver'
+  ]
+)
+
+python_tests(
+  name = 'artifact',
+  sources = ['test_artifact.py'],
+  dependencies = [
+    'src/python/pants/cache',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:dirutil',
   ]
 )
 
 python_tests(
   name = 'artifact_cache',
-  sources = globs('*.py'),
+  sources = ['test_artifact_cache.py'],
   dependencies = [
-    'src/python/pants/base:build_invalidator',
     'src/python/pants/cache',
+    'src/python/pants/invalidation',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test:base_test',
@@ -24,7 +35,7 @@ python_tests(
 
 python_tests(
   name = 'cache_setup',
-  sources = globs('*.py'),
+  sources = ['test_cache_setup.py'],
   dependencies = [
     'src/python/pants/backend/core/tasks:task',
     'src/python/pants/cache',
