+++ tests/python/pants_test/cache/test_artifact_cache.py
@@ -12,10 +12,10 @@ import unittest
 from contextlib import contextmanager
 from threading import Thread
 
-from pants.base.build_invalidator import CacheKey
 from pants.cache.artifact_cache import call_insert, call_use_cached_files
 from pants.cache.local_artifact_cache import LocalArtifactCache, TempLocalArtifactCache
 from pants.cache.restful_artifact_cache import InvalidRESTfulCacheProtoError, RESTfulArtifactCache
+from pants.invalidation.build_invalidator import CacheKey
 from pants.util.contextutil import pushd, temporary_dir, temporary_file
 from pants.util.dirutil import safe_mkdir
 from pants_test.base.context_utils import create_context
