+++ tests/python/pants_test/targets/test_sort_targets.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 import pytest
 
-from pants.base.build_graph import CycleException, sort_targets
+from pants.build_graph.build_graph import CycleException, sort_targets
 from pants_test.base_test import BaseTest
